--- Vacations.py
@@ -1,319 +1,0 @@
-direction = {
-    "Производство": {
-        "Череповец": "Специалист(по учету), Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Электросварщик ручной сварки, Череповец, без опыта, полная занятость, полный день, от 41 200 ₽\n\n\
-Инженер-технолог, фасонный профиль, Череповец, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Инженер-технолог КлЦ, Череповец, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Моторист самостоятельного управления судовым двигателем (рулевой) 5 разряд, Череповец, от 1 года до 3 лет, полная занятость, сменный график,40000 ₽\n\n\
-Слесарь-ремонтник, Череповец, от 1 года до 3 лет, полная занятость, сменный график, зарплата не указана\n\n\
-Электромонтер по ремонту и обслуживанию электрооборудования, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Мастер по ремонту оборудования (электрооборудование), Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Слесарь-ремонтник 4-6 разряд, Череповец, без опыта, полная занятость, полный день, 50000-70000 ₽\n\n\
-Токарь-расточник 5 разряд, Череповец, от 1 года до 3 лет, полная занятость, сменный график, зарплата не указана\n\n\
-Слесарь-ремонтник по ремонту гидрооборудования 5 разряд, Череповец, от 1 года до 3 лет, полная занятость, полный день, 40 000 - 50 000 ₽\n\n\
-Специалист ТОиР, Череповец, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Весовщик, Череповец, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Мастер по ремонту оборудования (электромеханик) ЦТОиР, Череповец, от 1 года до 3 лет, полная занятость, полный день, 88000 ₽\n\n\
-Мастер по ремонту оборудования, Череповец, от 1 года до 3 лет, полная занятость, полный день, 68000-78000 ₽\n\n\
-Специалист 1 категории, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Слесарь-ремонтник, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Токарь, Череповец, от 1 года до 3 лет, полная занятость, сменный график, зарплата не указана\n\n\
-Газорезчик, Череповец, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Электросварщик ручной сварки, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Арматурщик, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Монтажник оборудования металлургических заводов, Череповец, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Мастер по ремонту оборудования, Череповец, от 3 до 6 лет, полная занятость, полный день, до 85 000 ₽\n\n\
-Бригадир шихтового двора в сталеплавильном и ферросплавном производствах, Череповец, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Резчик холодного металла, Череповец, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Электросварщик ручной сварки 5 разряд, Череповец, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Машинист крана металлургического производства 4 разряд, Череповец, от 1 года до 3 лет, полная занятость, сменный график, 57000 ₽\n\n\
-Токарь 4-5 разряд, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Машинист крана 5 разряд, Череповец, без опыта, полная занятость, сменный график, зарплата не указана\n\n\
-Бригадир электромонтажников, Череповец, от 1 года до 3 лет, полная занятость, полный день, 40000 тысяч рублей ₽\n\n\
-Ведущий эксперт (развитие новых видов продукции), Череповец, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Электросварщик ручной сварки 5 разряд, Череповец, от 1 года до 3 лет, полная занятость, полный день, 45000 ₽\n\n\
-Слесарь-ремонтник, Череповец, от 1 года до 3 лет, полная занятость, сменный график, зарплата не указана\n\n\
-Слесарь по ремонту автомобилей, Череповец, от 1 года до 3 лет, полная занятость, сменный график, 37 500 ₽\n\n\
-Более детальная информация о вакансиях доступна по ссылке:https://clck.ru/UE67q",
-        "Москва": "Менеджер проектов в строительной отрасли, Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по работе с Исполнителями, Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Электромонтажник, Москва, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Руководитель порошкового и аддитивного направления, Москва, более 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по работе с заявками - КАМ, Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UYrhw",
-        "Санкт-Петербург": "Старший менеджер по перспективному инжинирингу, Санкт-Петербург, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Бригадир на отделке, сортировке, приемке, сдаче, пакетировке и упаковке металла и готовой продукции, Санкт-Петербург, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UZE5N",
-        "Колпино": "Ведущий проектировщик-технолог, Колпино, более 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Ведущий инженер-конструктор, Колпино, более 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Инженер–проектировщик (гидравлика), Колпино, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Инженер–проектировщик (Водоснабжение и канализация), Колпино, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Ведущий инженер-проектировщик (Теплоснабжение), Колпино, более 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Ведущий инженер-проектировщик (гидро- и пневмоуправление, смазочные системы), Колпино, более 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Ведущий инженер-проектировщик (газоснабжение), Колпино, более 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Ведущий инженер-проектировщик (вентиляция и аспирация), Колпино, более 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер проекта (толстый лист и трубы большого диаметра), Колпино, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по технологии производства труб, Колпино, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UZEDS",
-        "Воркута": 'Кладовщик, Воркута, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Участковый маркшейдер, Воркута, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Инженер-оператор аэрогазового контроля участка АБ(ВТБ) Шахта Заполярная, Воркута, без опыта, полная занятость, сменный график, зарплата не указана\n\n\
-Обучение по профессии "Электрослесарь дежурный по ремонту оборудования 3 разряда", Воркута, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по активам (работа в г.Воркута), Воркута, от 1 года до 3 лет, полная занятость, полный день, от 70000 (по результатам собеседования) ₽\n\n\
-Слесарь по ремонту автомобилей 3 разряда, Воркута, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UZEt7',
-        "Белгород": "Машинист погрузочно-доставочной машины, Белгород, от 1 года до 3 лет, полная занятость, сменный график, зарплата не указана\n\n\
-Горнорабочий на геологических работах, Белгород, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Горный мастер, Белгород, от 1 года до 3 лет, полная занятость, сменный график, зарплата не указана\n\n\
-Электрогазосварщик, Белгород, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Гидрогеолог, Белгород, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Машинист электровоза 4 разряда с полным рабочим днем на подземных работах, Белгород, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Горнорабочий подземный, Белгород, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Машинист горных выемочных машин, Белгород, от 1 года до 3 лет, полная занятость, сменный график, зарплата не указана\n\n\
-Машинист подземных самоходных машин, Белгород, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UZEx3",
-        "Яковлево": "Водитель автомобиля, занятый на транспортировании горной массы (МАЗ, КАМАЗ и т.п.), Яковлево, без опыта, полная занятость, сменный график, зарплата не указана\n\n\
-Доставщик крепежных материалов в шахту, занятый на поверхности, Яковлево, без опыта, полная занятость, сменный график, зарплата не указана\n\n\
-Подсобный рабочий, Яковлево, без опыта, полная занятость, сменный график, зарплата не указана\n\n\
-Геолог участковый, Яковлево, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Электрослесарь по ремонту и обслуживанию оборудования, Яковлево, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UZF35",
-        "Костомукша": "Машинист экскаватора 8 разряда, Костомукша, от 1 года до 3 лет, полная занятость, сменный график, 115000-120000 ₽\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UZF7K",
-        "Шексна": "Электромонтер по ремонту и обслуживанию электрооборудования, Шексна, без опыта, полная занятость, полный день, 45.000 ₽\n\n\
-Дефектоскопист по магнитному и ультразвуковому исследованию, Шексна, без опыта, полная занятость, полный день, от 40.000 до 50.000 ₽",
-        "Строитель": "Доставщик крепежных материалов в шахту 3 разряда, занятый на подземных работах 50% и более рабочего времени, Строитель, без опыта, полная занятость, сменный график, зарплата не указана",
-    },
-    "IT & Digital": {
-        "Колпино": "Ведущий специалист АСУ ТП, Колпино, более 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Инженер-проектировщик АСУ ТП, Колпино, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UZFH3",
-        "Москва": 'Специалист по информационной безопасности #Делетрон, Москва, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Сервис-менеджер (MES-системы), Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Администратор проекта, Москва, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Консультант Поддержки и развития функционала SAP APO, Москва, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Старший консультант, Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Системный архитектор, Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-PHP-разработчик, Москва, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Angular-разработчик, Москва, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Frontend-разработчик, Москва, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер (ИТ Консультант), Москва, более 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Бизнес-аналитик, Москва,, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Разработчик UI5, Москва, от 1 года до 3 лет, полная занятость, оолный день, зарплата не указана\n\n\
-Консультант SAP ЭДО, Москва, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Старший специалист по нагрузочному тестированию(Цифровая сталь), Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Старший специалист по функциональному тестированию(мобильные приложения), Москва, от 3 до 6 лет, полная занятость, полный день,зарплата не указана\n\n\
-Системный аналитик (e-com), Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Старший инженер по тестированию, Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Старший аналитик данных, Москва, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Бизнес-аналитик, Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Сервис-менеджер, Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Руководитель проектов, Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Старший консультант SAP MM, Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Консультант Power BI, Москва, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Консультант SAP BW, Москва, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Старший консультант SAP BW, Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Product manager/Senior Product manager (проект "Industrial Procurement Hub" ), Москва, более 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Старший консультант SAP TM, Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Тест-менеджер R2R(МСФО, Налоги, Базовый учет), Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Консультант SAP SD, Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Head of quality in IT, Москва, более 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UE77N',
-        "Белгород": "Специалист по автоматизации (АСУТП), Белгород, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UZFLt",
-        "Ярославль": "Младший специалист/младший консультант, Ярославль, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Ведущий специалист по обработке данных Excel, Ярославль, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Старший менеджер по информационной безопасности, Ярославль, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Стажер роботизации процессов, Ярославль, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Специалист роботизации процессов, Ярославль, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UZFyf",
-        "Череповец": "iOS-разработчик (удаленно), Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Администратор по поддержке инфраструктуры ЦОД, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Специалист (по автоматизации), Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Программист, Череповец, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Специалист группы разработки(направление отчетности), Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Младший консультант, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Администратор баз данных MSSQL, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Ведущий эксперт (автоматизация и цифровые технологии), Череповец, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Системный администратор Linux, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Ведущий сетевой инженер, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Специалист (разработка программного обеспечения), Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Консультант ЭДО (Kofax), Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Разработчик SAP IDeal, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Веб-разработчик Javascript (back-end Node.js, front-end React), Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Специалист (электроник), Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Ведущий эксперт (по автоматизации процессов ТОиР), Череповец, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Ведущий специалист, Череповец, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Специалист (предотвращение проблем), Череповец, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Младший специалист/Специалист, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Администратор Skype for Business, Череповец, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Администратор подсистемы доступа в Интернет, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Младший специалист\Специалист, Череповец, без опыта, полная занятость, сменный график, зарплата не указана\n\n\
-Младший специалист SAP APO, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Android разработчик, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Администратор MS AD, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Специалист по техническому обслуживанию систем неразрушающего контроля, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Специалист, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Ведущий специалист (1С), Череповец, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Системный администратор Trend Micro Deep Security, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Специалист группы разработки, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Администратор средств криптографической защиты информации, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер (1С), Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Специалист (1С), Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UE75r",
-        "Санкт-Петербург": "Fullstack разработчик, Санкт-Петербург, более 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Frontend разработчик, Санкт-Петербург, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Разработчик C# (чат-боты), Санкт-Петербург, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Руководитель проекта, Санкт-Петербург, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Консультант SAP MM, Санкт-Петербург, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Ведущий Full stack разработчик, Санкт-Петербург, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UZG5A",
-        "Воронеж": "Java Spring Middle разработчик (удаленно), Воронеж, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-React JS middle\senior разработчик (удаленно), Воронеж, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Старший специалист, Воронеж, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Тестировщик, Воронеж, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Разработчик JavaScript (удаленный формат), Воронеж, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Разработчик Java Spring Middle (удаленно), Воронеж, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Специалист (ТОиР), Воронеж, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Старший специалист функционального тестирования, Воронеж, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Стажер, Воронеж, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Системный администратор SAP Basis, Воронеж, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Ведущий специалист по функциональному тестированию направления FI Казначейство, Воронеж, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Разработчик баз данных SQL Sybase, Воронеж, от 3 до 6 лет, полная занятость, полный день,зарплата не указана\n\n\
-Младший тестировщик (QA Automation Engineer / Тестировщик), Воронеж, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Консультант ЭДО Open Text, Воронеж, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UZGBa",
-        "Костомукша": "Специалист АСУТП, Костомукша, от 1 года до 3 лет, полная занятость, полный день, от 70000 ₽\n\n\
-Разработчик MS SQL (BI), Костомукша, от 1 года до 3 лет, полная занятость, полный день, от 69000 ₽\n\n\
-Разработчик Python, c# (IoT, CV, ML, AI), Костомукша, от 1 года до 3 лет, полная занятость, полный день, от 69000 ₽\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UZGNV",
-    },
-    "Офис": {
-        "Москва": "Аналитик HR процессов, Москва, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Инженер-пректировщик (Civil Engineer), Москва, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по продажам, Москва, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Агент по продажам, Москва, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по тендерам, Москва, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по продажам, Москва, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Кладовщик, Москва, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Старший менеджер (эксперт), Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по продажам, Москва, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Ведущий специалист по введению первичной бухгалтерской отчетности, Москва, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по финансовому контроллингу, Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по продвижению, Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по продвижению, Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по интернет продвижению, Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Senior Investment Manager/Руководитель проектов, Москва, более 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по МСФО, Москва, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер программы обучения по охране труда и промышленной безопасности, Москва, более 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Главный операционный маркетолог-аналитик, Москва, более 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Аналитик клиентского опыта, Москва, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UZGRE",
-        "Череповец": "Менеджер по снабжению (инвестиционные проекты), Череповец, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер/специалист по продажам, Калибровочное направление , Метиз, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер (навигатор), Череповец, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Электромеханик - 1 помощник механика по электрооборудованию, Череповец, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Специалист по закупкам, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Главный специалист (консолидированный анализ), Череповец, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Мастер, Череповец, от 1 года до 3 лет, полная занятость, сменный график, зарплата не указана\n\n\
-Менеджер (аналитик), Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер, Череповец, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер-экономист по контроллингу, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Маркетолог-аналитик, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Старший менеджер (Бизнес-система «Северстали»), Череповец, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по строительному контролю механомонтажных работ, Череповец, более 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по строительному контролю электромонтажных работ, Череповец, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Специалист (экономист по контроллингу), Череповец, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по строительному контролю электромонтажных работ, Череповец, более 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по строительному контролю механомонтажных работ, Череповец, более 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Главный специалист КМ (конструкции металлические), Череповец, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Специалист, Череповец, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Специалист (конструктор) 1 кат., Череповец, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Специалист по планированию закупок, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по строительному контролю общестроительных работ, Череповец, более 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по совершенствованию производственного процесса, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер группы продаж на экспорт. Северсталь Канаты, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по организации отдыха, Череповец, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Повар, Череповец, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Специалист по геодезии, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Специалист по оперативной работе, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по флоту (заместитель начальника цеха по обеспечению безопасности судоходства), Череповец, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по планированию, Череповец, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Старший менеджер офиса поддержки экспертной сети, Череповец, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Старший юрисконсульт, Череповец, более 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UEBKJ",
-        "Санкт-Петербург": "Специалист по продажам, Санкт-Петербург, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер Фрахтового Центра, Санкт-Петербург, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Специалист управления охраны труда и промышленной безопасности г. Колпино (Санкт-Петербург), Колпино, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по продажам, Санкт-Петербург, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по продажам, Санкт-Петербург, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Специалист-стажер (на разбор архива), Санкт-Петербург, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по продвижению на экспортных рынках, Санкт-Петербург, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UZGU5",
-        "Самара": "Специалист по продажам, Самара, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по продажам, Самара, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана",
-        "Нижний Новгород": "Менеджер по продажам, Нижний Новгород, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана",
-        "Мурманск": "Менеджер по продажам, Мурманск, без опыта, полная занятость, полный день, зарплата не указана",
-        "Ярославль": "Специалист по формированию бухгалтерской, налоговой и статистической отчетности, Ярославль, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Ведущий специалист по казначейским операциям со знанием английского языка, Ярославль, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Специалист по учету дебиторской задолженности, Ярославль, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Стажер по дебиторской задолженности, Ярославль, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер Центра экспертизы по учету основных средств, Ярославль, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Ведущий специалист Центра экспертизы, Ярославль, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Специалист (нецентрализованные закупки), Ярославль, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Ведущий специалист по формированию бухгалтерской отчетности, Ярославль, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Специалист по договорам страхования, Ярославль, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Ведущий специалист Кадры ЦЕС, Ярославль, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Товаровед, Ярославль, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Ведущий специалист по формированию бухгалтерской отчетности по РСБУ/МСФО, Ярославль, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Специалист по банковским операциям, Ярославль, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Специалист по работе с казначейскими документами, Ярославль, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Специалист по ведению бухгалтерского учета, Ярославль, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Специалист по финансовым сделкам, Ярославль, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Ведущий специалист по закупкам услуг, Ярославль, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Стажер по работе с запросами, Ярославль, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Администратор проекта, Ярославль, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Специалист по проведению инвентаризации, Ярославль, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Специалист по расчётам с персоналом, Ярославль, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UZGiT",
-        "Колпино": "Инженер–проектировщик (Силовое электрооборудование), Колпино, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по развитию продуктовой категории, Колпино, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер (по совершенствованию производственного процесса, трубный прокат), Колпино, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер службы обеспечения бизнеса, Колпино, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Ведущий специалист КЖ (строительный отдел), Колпино, более 6 лет, полная занятость, полный день, 80000 ₽\n\n\
-Менеджер по техническим продажам, Колпино, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UZGo8",
-        "Казань": "Специалист-стажер в г. Казань, Казань, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Специалист по продажам, Казань, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по продажам, Казань, от 3 до 6 лет, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UZGuN",
-        "Набережные Челны": "Менеджер по продажам, Набережные Челны, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана",
-        "Батайск": "Специалист по продажам, Батайск, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Менеджер по продажам, Батайск, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UZGzR",
-        "Подольск": "Специалист по транспорту, Подольск, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана",
-        "Моссан-лез-Альпий": "Тест, Моссан-лез-Альпий, без опыта, полная занятость, полный день, зарплата не указана",
-        "Волгоград": "Аналитик-маркетолог, Волгоград, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана",
-        "Костомукша": "Машинист тепловоза, тягового агрегата, Костомукша, от 1 года до 3 лет, полная занятость, сменный график, от 70 000 до 90 000 ₽\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UZH8v",
-        "Воронеж": "Менеджер по продажам, Воронеж, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UZHGS",
-        "Красный Сулин": "Товаровед, Красный Сулин, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: ",
-        "Ступино": "Специалист в Ступино, Ступино, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: ",
-        "Тверь": "Специалист по сопровождению продаж, Тверь, от 1 года до 3 лет, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: ",
-    },
-    "Молодым специалистам": {
-        "Череповец": "Специалист-стажер в Группу сопровождения проектных закупок, Череповец, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Практика для студентов, Череповец, без опыта, полная занятость, полный день, 20 000 ₽\n\n\
-Стажировка для выпускников, Череповец, без опыта, полная занятость, полный день, 24 000 ₽\n\n\
-Стажер в Дирекцию по техническому развитию и качеству, Череповец, без опыта, полная занятость, полный день\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UZHLU",
-        "Ярославль": "Стажер по формированию бухгалтерской, налоговой и статистической отчетности, Ярославль, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Стажер по формированию бухгалтерской отчетности, Ярославль, без опыта, полная занятость, полный день, зарплата не указана\n\n\
-Более детальная информация о вакансиях доступна по ссылке: https://clck.ru/UZHQS",
-}
-}
