--- history.py
@@ -1,37 +1,0 @@
-def history_of_severstal():
-    history = "24 августа 1955 года был получен первый череповецкий чугун на доменной печи №1. Этот день стал Днем рождения Череповецкого металлургического завода.\n\n\
-1956 – запущена первая коксовая батарея.\n\n\
-1956 – задута ДП №2.\n\n\
-1958 – получен первый слиток мартеновской стали.\n\n\
-1959 – введен в эксплуатацию первый прокатный стан 1150.\n\n\
-1962 – задута ДП №3.\n\n\
-1969 – получена первая электросталь.\n\n\
-1969 – задута ДП №4.\n\n\
-1975 – введен в эксплуатацию стан 2000.\n\n\
-1980 – получена первая конверторная сталь.\n\n\
-1986 – задута ДП №5.\n\n\
-1993 -  Череповецкий металлургический комбинат зарегистрирован в мэрии Череповца как акционерное общество «Северсталь».\n\n\
-1999 – «Северсталь» объявила о приобретении АО «Карельский окатыш».\n\n\
-2001 – «Северсталь» объявила о приобретении АО «Олкон».\n\n\
-2003 – «Северсталь» объявила о приобретении АО «Воркутауголь».\n\n\
-2004 – «Северсталь» объявила о приобретении компании Rouge Steel, которая впоследствии была переименована в Severstal Dearborn.\n\n\
-2004 – образование компании «Северсталь Метиз».\n\n\
-2006 – старт торговли GDR на акции «Северстали» на Лондонской фондовой бирже.\n\n\
-2006 – введен в эксплуатацию Ижорский трубный завод.\n\n\
-2008 – «Северсталь» объявила о приобретении активов в США (Severstal Warren, Severstal Wheeling и Severstal Sparrows Point).'\n\n\
-2008 – «Северсталь» объявила о приобретении канадской компании PBS Coals Corporation.'\n\n\
-2010 – введен в эксплуатацию трубопрофильный завод «Северсталь ТПЗ – Шексна».\n\n\
-2011 - «Северсталь» объявила о продаже трёх активов в США (Severstal Warren, Severstal Wheeling и Severstal Sparrows Point).\n\n\
-2012 – выделение золотодобывающих активов «Северстали» в компанию Nordgold.\n\n\
-2013 – создание дистрибуционной сети дивизиона «Северсталь Российская сталь».\n\n\
-2014 – введен в эксплуатацию сортовой завод «Северсталь – Сортовой завод Балаково».\n\n\
-2014 — шахта «Воркутинская» АО «Воркутауголь» запустила лаву-гигант — длиной 317 м, с запасами более 1,4 млн тонн угля и мощностью пласта более трёх м.\n\n\
-2014 – «Северсталь» объявила о продаже Severstal Dearborn, Severstal Columbus и PBS Coals.\n\n\
-2017 – запуск интернет-магазина «Северстали».\n\n\
-2017 – «Северсталь» объявила о приобретении Яковлевского рудника.\n\n\
-2017 – образование «Северсталь Диджитал» для внедрения цифровой стратегии компании.\n\n\
-2017 – создание крупнейшего среди российских промышленных предприятий data lake.\n\n\
-2018 – «Северсталь», РОСНАНО и Windar Renovables открыли первое в России производство башен ветроустановок.\n\n\
-2018 – на Дне инвестора «Северсталь» объявила о новой стратегии компании и планах ежегодного увеличения EBITDA в среднем на 10 - 15%. Для реализации поставленных задач в «Северстали» запущена программа трансформации.\n\n\
-2019 – «Северсталь» и Tenaris объявили о создании совместного предприятия по производству сварных труб в Западной Сибири."
-    return history
