+++ persistence/manufacture_parser.py
