--- requirements.txt
@@ -1,2 +1,0 @@
-aiogram
-investpy
